Natwende2 initial commit

+++ ControlModule/natwende-mukwai-web/src/main/java/com/mweka/natwende/helper/ApplicationBean.java
@@ -25,7 +25,8 @@ import javax.inject.Named;
 import javax.servlet.ServletContext;
 
 import com.mweka.natwende.resource.elasticdb.ElasticDB;
-import com.mweka.natwende.route.vo.StopVO;
+import com.mweka.natwende.route.vo.RouteStopLinkVO;
+import com.mweka.natwende.route.vo.RouteVO;
 import com.mweka.natwende.trip.vo.ElasticTripVO;
 import com.mweka.natwende.trip.vo.TripVO;
 import com.mweka.natwende.util.MapObjectInstance;
@@ -49,8 +50,9 @@ public class ApplicationBean {
 	@PostConstruct
 	public void init(){
 		applicationProperties = loadManifestFile();
-		loadActiveTrips();
 		cacheMap = new ConcurrentHashMap<>();
+		loadStationsAndOrderOfIndexesPerRoute();
+		loadActiveTrips();		
 	}
 	
 	public void init(@Observes @Initialized(ApplicationScoped.class) Object init) {
@@ -80,22 +82,54 @@ public class ApplicationBean {
 	public void loadActiveTrips() {
 		try {
 			List<TripVO> resultList = serviceLocator.getTripDataFacade().getActiveTrips();
-			ElasticTripVO elasticData = new ElasticTripVO();
-			Map<String, Object> data = null;			
+			if (resultList.isEmpty()) {
+				return; // Warn and return
+			}
+			
+			@SuppressWarnings("unchecked")
+			List<RouteVO> routeList = (List<RouteVO>) cacheMap.get("ROUTES_WITH_LINKED_STOPS_INCLUDING_FIRST_STOP_AND_FINAL_DESTINATION");
+			
 			for (TripVO trip : resultList) {				
 				elasticDB.deleteEntity(elasticDB.getEntity(trip.getUniqueId()));				
-				elasticData = TripVO.convertToElasticData(trip, elasticData);
-				List<StopVO> stationList = serviceLocator.getStopDataFacade().getByRoute(trip.getTripSchedule().getRoute());
-				for (StopVO station : stationList) {
-					elasticData.getStationList().add(station.getTown().getDisplay());
+				ElasticTripVO elasticData = TripVO.convertToElasticData(trip, new ElasticTripVO());
+				//List<StopVO> stationList = serviceLocator.getStopDataFacade().getByRoute(trip.getTripSchedule().getRoute());
+				
+				RouteVO targetRoute = trip.getTripSchedule().getRoute();
+				if (targetRoute.getRouteStops().isEmpty()) {
+					for (RouteVO route : routeList) {
+						if (route.equals(targetRoute)) {
+							targetRoute.setRouteStops(route.getRouteStops());
+							break;
+						}
+					}
 				}
-				data = MapObjectInstance.parameters(elasticData);
+				if (targetRoute.getRouteStops().isEmpty()) {
+					return; // Warn and return
+				}
+				
+				for (int j = 0; j < targetRoute.getRouteStops().size(); j++) {					
+					RouteStopLinkVO stationLink = targetRoute.getRouteStops().get(j);
+					System.out.println(stationLink);
+					elasticData.getStationList().add(stationLink.getStationIndex() + "_" + stationLink.getStop().getTown().getDisplay());
+				}
+				System.out.println("====================================================================");
+				Map<String, Object> data = MapObjectInstance.parameters(elasticData);
 				elasticDB.insertEntity(data);
 			}
 		}
 		catch (Exception ex) {
 			LOGGER.log(Level.SEVERE, "Exception", ex);
 		}
+	}
+	
+	public void loadStationsAndOrderOfIndexesPerRoute() {
+		List<RouteVO> allRoutes = serviceLocator.getRouteDataFacade().getAll();		
+		for (RouteVO route : allRoutes) {
+			route.getRouteStops().add(new RouteStopLinkVO(-1, route, route.getStart()));
+			route.getRouteStops().addAll(serviceLocator.getRouteStopLinkDataFacade().getAllByRoute(route));
+			route.getRouteStops().add(new RouteStopLinkVO(route.getRouteStops().size() - 1, route, route.getStop()));			
+		}
+		cacheMap.put("ROUTES_WITH_LINKED_STOPS_INCLUDING_FIRST_STOP_AND_FINAL_DESTINATION", allRoutes);
 	}
 
 	public Properties getApplicationProperties() {
